reverted old commits and new gadget file implemented
